pyproject.toml
[project]
name = "agent-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=1.28.0",
    "python-dotenv>=1.0.1",
]
